requirements.txt
Flask==2.3.2
openai==0.27.4
gunicorn==20.1.0
python-dotenv==1.0.0
requests==2.31.0